Process queued service discovery only when link is connected
Service discovery requests for BLE and BR/EDR links are queued
when there is a service discovery already in progress.

When a link gets disconnected while processing a queued service
discovery, the service discovery gets ignored and the state
machine is not moved to idle. Hence the subsequent service
discovery requests are not processed by the stack.

This change checks whether the link is connected before
processing the queued service discovery requests.

Change-Id: Idfcdfd6a9bdd12d6ca71e6c9caa72d2df32566c7